rename: throttle 훅 이름 변경, 파일 이동
